settings: 拆出 MpcSettingsController（双 sub-controller 第 2 步，完成）
将 SettingsController 剩余的 ~50 个 MPC 方法、14 个 mpc 状态字段、mpc 的 DOM 事件
绑定整体拆到 js/controller/settings/mpc-settings-controller.js；DEFAULT_MPC_* 常量
补入 settings/settings-utils.js。

- 主 class 重写为只保留授权站点 / 改密 / 重置钱包 + 持有两个 sub-controller 实例，
  bindEvents 仅做 backup/mpc 委托 + 自身绑定；loadMpcSettings/loadMpcSessions 委托。
- 一并删除随 backup/mpc 拆走后不再使用的共享 helper（normalize*/getDefault*/
  isLegacy*/deriveUcanAudience）与多余 import。
- 主 class 3077 → 512 行（backup 1211 + mpc 1326 + utils 118 三个文件分担）。
- popup-controller 公开 API 与调用点零改动；已做模块加载 + 实例化 + bindEvents 冒烟、
  残留引用扫描；MPC tab 交互需在 Chrome 实测。
